ref: Introduce base classes for coordinate systems
