+++ src/Math/Calculator/Base/AbstractCalculator.php
@@ -1,10 +1,15 @@
 <?php
 
-namespace Famoser\Elliptic\Math\Calculator;
+namespace Famoser\Elliptic\Math\Calculator\Base;
 
 use Famoser\Elliptic\Primitives\Curve;
 
-abstract class BaseCalculator
+/**
+ * @template T
+ * 
+ * @implements CalculatorInterface<T>
+ */
+abstract class AbstractCalculator implements CalculatorInterface
 {
     public function __construct(private readonly Curve $curve)
     {
